+++ tests/unit/datazip_test.py
@@ -43,6 +43,7 @@ def test_datazip(temp_dir):
             z.writed("b", df_dict["b"])
             z.writed("c", {1: 3, "3": "fifteen", 5: (0, 1)})
             z.writed("aa", df_dict["a"].loc[:, (0, "a")])
+            z.writed("aaa", df_dict["a"].loc[:, [(0, "a")]])
             z.writed("d", "hello world")
             with pytest.raises(FileExistsError):
                 z.writed("c", {1: 3, "3": "fifteen", 5: (0, 1)})
@@ -58,6 +59,9 @@ def test_datazip(temp_dir):
             assert isinstance(z1.read("a"), pd.DataFrame)
             assert isinstance(z1.read("b"), pd.Series)
             assert isinstance(z1.read("aa"), pd.Series)
+            aaa = z1.read("aaa")
+            assert isinstance(aaa, pd.DataFrame)
+            assert isinstance(aaa.columns, pd.MultiIndex)
             assert isinstance(z1.read("c"), dict)
             assert isinstance(z1.read("d"), str)
             assert z1.read("d") == "hello world"